Move to a src layout
Store the source code in `src` and the tests in `test` outside the
package. This was the tests aren't packaged, saving a bit of storage of
the package. Also simplifies out testing in the Makefile.
