--- ensaio/tests/__init__.py
@@ -1,6 +1,0 @@
-# Copyright (c) 2021 The Ensaio Developers.
-# Distributed under the terms of the BSD 3-Clause License.
-# SPDX-License-Identifier: BSD-3-Clause
-#
-# This code is part of the Fatiando a Terra project (https://www.fatiando.org)
-#
